requirements.txt
python-telegram-bot>=20.0
gspread>=6.0.0
google-auth>=2.0.0
requests>=2.31.0
python-dotenv>=1.0.0
